code/requirements.txt
pyyaml
apscheduler
confluent-kafka
polars
prometheus-client
elasticsearch
